Поменял accrual uint32 => float32

+++ internal/app/accrualservice/models.go
@@ -2,7 +2,7 @@ package accrualservice
 
 type GetOrderResponse struct {
 	// Номер заказа
-	Order   uint32  `json:"order"`
-	Status  string  `json:"status"`
-	Accrual *uint32 `json:"accrual,omitempty"`
+	Order   uint32   `json:"order"`
+	Status  string   `json:"status"`
+	Accrual *float32 `json:"accrual,omitempty"`
 }
