+++ internal/app/gophermartprovider/models/order.go
@@ -1,8 +1,8 @@
 package models
 
 type Order struct {
-	Number     string  `db:"number"`
-	Status     string  `db:"status"`
-	Accrual    *uint32 `db:"accrual,omitempty"`
-	UploadedAt string  `db:"uploaded_at"`
+	Number     string   `db:"number"`
+	Status     string   `db:"status"`
+	Accrual    *float32 `db:"accrual,omitempty"`
+	UploadedAt string   `db:"uploaded_at"`
 }
