+++ internal/app/usecase/order/models/order.go
@@ -1,8 +1,8 @@
 package models
 
 type Order struct {
-	Number     string  `json:"number"`
-	Status     string  `json:"status"`
-	Accrual    *uint32 `json:"accrual"`
-	UploadedAt string  `json:"uploaded_at"`
+	Number     string   `json:"number"`
+	Status     string   `json:"status"`
+	Accrual    *float32 `json:"accrual"`
+	UploadedAt string   `json:"uploaded_at"`
 }
